+++ MakemoreC/value.cpp
@@ -1,12 +1,12 @@
 #include "pch.h"
 #include <value.h>
 
-value operator+(float f, value& v)
+value operator+(double f, value& v)
 {
 	return v + f;
 }
 
-value operator*(float f, value& v)
+value operator*(double f, value& v)
 {
 	return v * f;
 }
